Finalizando Services e DTO

+++ src/main/java/com/guilherme/aulaspring/projetofinalsalaocarmemlucia/model/Servico.java
@@ -3,6 +3,8 @@ package com.guilherme.aulaspring.projetofinalsalaocarmemlucia.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name="servico")
 @Data
@@ -15,14 +17,14 @@ public class Servico {
     @Column(name="nome", length = 100, nullable = false)
     private String nome;
     @Column(name="valor", length = 10, nullable = false)
-    private Double valor;
+    private BigDecimal valor;
     @Column(name="temposervico")
     private Integer duracaoMinutos;
 
     public Servico() {
     }
 
-    public Servico(Long id, String nome, Double valor, Integer duracaoMinutos) {
+    public Servico(Long id, String nome, BigDecimal valor, Integer duracaoMinutos) {
         this.id = id;
         this.nome = nome;
         this.valor = valor;
